separated files by hardware + migrated library files to src
