+++ src/Lantern.h
